Second step commit: getDeviceDescriptor (C structure)

+++ src/ru/shurikvo/jlibusb/DeviceDescriptor.java
@@ -3,14 +3,14 @@ package ru.shurikvo.jlibusb;
 public class DeviceDescriptor {
     public byte bLength;
     public byte bDescriptorType;
-    public byte bcdUSB;
+    public short bcdUSB;
     public byte bDeviceClass;
     public byte bDeviceSubClass;
     public byte bDeviceProtocol;
     public byte bMaxPacketSize0;
-    public byte idVendor;
-    public byte idProduct;
-    public byte bcdDevice;
+    public short idVendor;
+    public short idProduct;
+    public short bcdDevice;
     public byte iManufacturer;
     public byte iProduct;
     public byte iSerialNumber;
